Build : api 키 숨기기, 위치관련 컴포너트 분리
