+++ app/build.gradle
@@ -4,7 +4,8 @@ plugins {
     id 'kotlin-kapt'
     id 'dagger.hilt.android.plugin'
 }
-
+Properties properties = new Properties()
+properties.load(project.rootProject.file("local.properties").newDataInputStream())
 android {
     namespace 'com.pat.miraclepat'
     compileSdk 34
@@ -15,6 +16,9 @@ android {
         targetSdk 34
         versionCode 1
         versionName "1.0"
+
+        buildConfigField "String", "CLIENT_ID", properties["CLIENT_ID"]
+        buildConfigField "String", "CLIENT_SECRET", properties["CLIENT_SECRET"]
 
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
@@ -35,6 +39,7 @@ android {
     buildFeatures {
         viewBinding true
         dataBinding true
+        buildConfig true
     }
 }
